pyproject.toml
[tool.poetry]
name = "holosiivska_invest_bot"
version = "0.1.0"
description = "Telegram bot for Investors of Residential Complex 'Holosiivska Dolyna'"
authors = ["Dmytro Chystiakov <dlchistyakov@gamil.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0"
python-dotenv = "^1.0.1"
pytest-asyncio = "^0.23.8"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"